chore: Single initUserStore call
