+++ apps/editor.planx.uk/src/routes/(auth)/login.tsx
@@ -1,7 +1,5 @@
-import { createFileRoute, redirect } from "@tanstack/react-router";
+import { createFileRoute } from "@tanstack/react-router";
 import DelayedLoadingIndicator from "components/DelayedLoadingIndicator/DelayedLoadingIndicator";
-import { getCookie } from "lib/cookie";
-import { useStore } from "pages/FlowEditor/lib/store";
 import Login from "pages/Login/Login";
 import { z } from "zod";
 
@@ -11,38 +9,6 @@ const loginSearchSchema = z.object({
 
 export const Route = createFileRoute("/(auth)/login")({
   validateSearch: loginSearchSchema,
-
-  beforeLoad: async ({ search }) => {
-    // Check for auth cookie
-    const authCookie = getCookie("auth");
-    if (!authCookie) {
-      return {};
-    }
-
-    // User is already authenticated - determine where to redirect
-    // If there's an explicit redirectTo to a specific page, handle it
-    if (search.redirectTo && search.redirectTo !== "/") {
-      throw redirect({ to: search.redirectTo });
-    }
-
-    // Otherwise, redirect to the user's default team if they have one
-    const user = await useStore.getState().initUserStore();
-    if (user?.defaultTeamId) {
-      const defaultTeam = user.teams.find(
-        (t) => t.team.id === user.defaultTeamId,
-      );
-      if (defaultTeam) {
-        throw redirect({
-          to: "/app/$team",
-          params: { team: defaultTeam.team.slug },
-        });
-      }
-    }
-
-    // If neither, redirect to `/app`
-    throw redirect({ to: "/app" });
-  },
-
   component: Login,
   pendingComponent: DelayedLoadingIndicator,
 });
